+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Post v1 a change is further considered breaking if the state file format is inco
 
 * Added a game variant message at the start of a game. ([#235](https://github.com/hjwylde/werewolf/issues/235))
 * Added the Saint role. ([#178](https://github.com/hjwylde/werewolf/issues/178))
+* Added marks to the Dullahan's status command. ([#236](https://github.com/hjwylde/werewolf/issues/236))
 
 ### v1.3.0.1